requirements.txt
Flask>=2.0.0,<3.0.0
Flask-SQLAlchemy>=3.0.0,<4.0.0
Flask-Migrate==4.0.5
marshmallow==3.20.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.0.2
Werkzeug==3.0.1
SQLAlchemy>=2.0.27,<3.0.0
sqlalchemy-utils==0.41.1
alembic==1.13.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
pluggy==1.4.0
typing_extensions==4.9.0 
sqlalchemy-utils==0.41.1